refactor: Product 클래스 만들어 VendingMachine 역할 분리하기 (#39)
